+++ packages/api-gateway/__tests__/geolocation.test.js
@@ -21,14 +21,14 @@ describe("geolocation/places endpoint", () => {
         }),
       ])
     );
-  });
+  }, 10000);
 
   test("Given wrong or null latitude, longitude or radius when a user select a mark in a map, then return status code 400 ", async () => {
-    const address = await api.inject({
+    const places = await api.inject({
       method: "GET",
       url: "geolocation/places",
     });
 
-    expect(address.statusCode).toEqual(400);
-  });
+    expect(places.statusCode).toEqual(400);
+  }, 10000);
 });
